MNT: Move set_and_wait to utils
