+++ ophyd/controls/device.py
@@ -4,7 +4,7 @@ import logging
 from collections import (OrderedDict, namedtuple)
 
 from .ophydobj import (OphydObject, DeviceStatus)
-from ..utils import TimeoutError
+from ..utils import TimeoutError, set_and_wait
 
 logger = logging.getLogger(__name__)
 
@@ -235,17 +235,6 @@ class ComponentMeta(type):
 
         return clsobj
 
-
-def set_and_wait(signal, val):
-    """
-    Set a signal to a value and wait until it reads correctly.
-
-    There are cases where this would not work well, so it should be revisited.
-    """
-    signal.put(val)
-    while signal.get() != val:
-        ttime.sleep(0.1)
-        logger.info("Waiting for %s to be set...", signal.name)
 
 
 # These stub 'Interface' classes are the apex of the mro heirarchy for
