+++ ophyd/utils/epics_pvs.py
@@ -336,3 +336,16 @@ def raise_if_disconnected(fcn):
         else:
             raise DisconnectedError('{} is not connected'.format(self.name))
     return wrapper
+
+
+def set_and_wait(signal, val):
+    """
+    Set a signal to a value and wait until it reads correctly.
+
+    There are cases where this would not work well, so it should be revisited.
+    """
+    signal.put(val)
+    while signal.get() != val:
+        ttime.sleep(0.1)
+        logger.info("Waiting for %s to be set...", signal.name)
+
